+++ Magic/docs/release-notes.md
@@ -3,6 +3,7 @@
 # Release notes
 ## Upcoming release
 * Added API methods to support Skill Prestige.
+* You now get 100 mana points immediately after the Wizard event, so you can start using magic on the same day.
 * The Wizard event is now skippable.
 * Fixed Wizard event broken if another mod changes the location in an incompatible way.
 * Fixed magic UI rendered before learning magic if another mod added mana points.